backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.18
httpx>=0.28.0
sse-starlette>=2.0.0
lxml>=5.3.0
Pillow>=10.0
numpy>=1.26.0
rembg>=2.0.50
onnxruntime>=1.17.0
pyyaml>=6.0
tqdm>=4.66.0
defusedxml>=0.7.1
pytest>=8.3.0
pytest-asyncio>=0.24.0
